Improve board loading behavior and normalize redux state
Instead of fetching a board, then it’s lists, then it’s cards, we fetch the board with it’s lists and cards, and then normalize all that data into redux.  To support this, I replaced the separate boards, lists, and cards reducers with a single entities reducer, and updated the related actions and selectors to work with the new reducer.

In the future I plan to move users, shares, and card details into the entities reducer as well.